+++ packages/@aws-cdk/cx-api/lib/index.ts
@@ -4,3 +4,4 @@ export * from './context/hosted-zone';
 export * from './context/vpc';
 export * from './context/ssm-parameter';
 export * from './context/availability-zones';
+export * from './metadata/assets';
