+++ packages/aws-cdk/integ-tests/test-cdk-synth.sh
@@ -10,6 +10,8 @@ assert "cdk synth cdk-toolkit-integration-test-1" <<HERE
 Resources:
   topic69831491:
     Type: AWS::SNS::Topic
+    Metadata:
+      aws:cdk:path: cdk-toolkit-integration-test-1/topic/Resource
 
 HERE
 
@@ -17,8 +19,12 @@ assert "cdk synth cdk-toolkit-integration-test-2" <<HERE
 Resources:
   topic152D84A37:
     Type: AWS::SNS::Topic
+    Metadata:
+      aws:cdk:path: cdk-toolkit-integration-test-2/topic1/Resource
   topic2A4FB547F:
     Type: AWS::SNS::Topic
+    Metadata:
+      aws:cdk:path: cdk-toolkit-integration-test-2/topic2/Resource
 
 HERE
