+++ packages/aws-cdk/test/test.cdk-doctor.ts
@@ -1,5 +1,6 @@
 import mockery = require('mockery');
 import { ICallbackFunction, Test, testCase } from 'nodeunit';
+import { CommandHandler } from '../lib/command-api';
 
 module.exports = testCase({
   '`cdk doctor`': {
@@ -16,8 +17,12 @@ module.exports = testCase({
       cb();
     },
     async 'exits with 0 when everything is OK'(test: Test) {
+      const argv: any = {};
       try {
-        const result = await require('../lib/commands/doctor').handler();
+        require('../lib/commands/doctor').handler(argv);
+        const result = await (argv.commandHandler as CommandHandler)({
+          args: argv
+        } as any);
         test.equal(result, 0, 'exit status was 0');
       } catch (e) {
         test.doesNotThrow(() => e);
